抽离错误判断逻辑
